requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.1
celery==5.5.3
redis==5.2.1
bcrypt==4.3.0
requests==2.31.0
openai==1.3.0
python-docx==1.1.0
PyPDF2==3.0.1
# Optional AI dependencies (install separately if needed)
# transformers==4.36.2
# torch==2.2.0
# sentence-transformers==2.2.2 